+++ toqito/channel_props/__init__.py
@@ -2,3 +2,4 @@
 from toqito.channel_props.is_herm_preserving import is_herm_preserving
 from toqito.channel_props.is_completely_positive import is_completely_positive
 from toqito.channel_props.is_positive import is_positive
+from toqito.channel_props.is_unital import is_unital